move diff repository tests around
