+++ backend/tests/unit/core/diff/repository/test_diff_repository.py
@@ -13,7 +13,6 @@ from infrahub.core.constants import DiffAction
 from infrahub.core.constants.database import DatabaseEdgeType
 from infrahub.core.diff.model.path import (
     BranchTrackingId,
-    EnrichedDiffNode,
     EnrichedDiffRoot,
     EnrichedDiffs,
     NameTrackingId,
@@ -22,11 +21,10 @@ from infrahub.core.diff.model.path import (
 from infrahub.core.diff.repository.deserializer import EnrichedDiffDeserializer
 from infrahub.core.diff.repository.repository import DiffRepository
 from infrahub.core.timestamp import Timestamp
-from infrahub.core.utils import delete_all_nodes
 from infrahub.database import InfrahubDatabase
 from infrahub.exceptions import ResourceNotFoundError
 
-from .factories import (
+from ..factories import (
     EnrichedAttributeFactory,
     EnrichedConflictFactory,
     EnrichedNodeFactory,
@@ -35,12 +33,14 @@ from .factories import (
     EnrichedRelationshipGroupFactory,
     EnrichedRootFactory,
 )
+from .base import DiffRepositoryTestBase
 
 
-class TestDiffRepositorySaveAndLoad:
-    @pytest.fixture
-    async def reset_database(self, db: InfrahubDatabase, default_branch):
-        await delete_all_nodes(db=db)
+class TestDiffRepositorySaveAndLoad(DiffRepositoryTestBase):
+    base_branch_name = "main"
+    diff_branch_name = "diff"
+    diff_from_time = DateTime.create(2024, 6, 15, 18, 35, 20, tz=UTC)
+    diff_to_time = DateTime.create(2024, 6, 15, 18, 49, 40, tz=UTC)
 
     @pytest.fixture
     def diff_repository(self, db: InfrahubDatabase) -> Generator[DiffRepository, None, None]:
@@ -53,86 +53,6 @@ class TestDiffRepositorySaveAndLoad:
         yield diff_repository
         config.SETTINGS.database.max_depth_search_hierarchy = original_depth
         config.SETTINGS.database.query_size_limit = original_size
-
-    def build_diff_node(self, num_sub_fields=2, no_recurse=False) -> EnrichedDiffNode:
-        enriched_node = EnrichedNodeFactory.build(
-            attributes={
-                EnrichedAttributeFactory.build(
-                    properties={
-                        EnrichedPropertyFactory.build(property_type=det)
-                        for det in random.sample(list(DatabaseEdgeType), num_sub_fields)
-                    }
-                )
-                for _ in range(num_sub_fields)
-            },
-            relationships={
-                EnrichedRelationshipGroupFactory.build(
-                    relationships={
-                        EnrichedRelationshipElementFactory.build(
-                            properties={
-                                EnrichedPropertyFactory.build(property_type=det)
-                                for det in random.sample(list(DatabaseEdgeType), num_sub_fields)
-                            }
-                        )
-                        for _ in range(num_sub_fields)
-                    },
-                    nodes=set(),
-                )
-                for _ in range(num_sub_fields)
-            },
-        )
-        if no_recurse:
-            return enriched_node
-        if num_sub_fields > 1 and len(enriched_node.relationships) > 0:
-            for relationship_group in enriched_node.relationships:
-                relationship_group.nodes = {
-                    self.build_diff_node(num_sub_fields=num_sub_fields - 1) for _ in range(num_sub_fields - 1)
-                }
-                break
-        return enriched_node
-
-    def setup_method(self):
-        self.base_branch_name = "main"
-        self.diff_branch_name = "diff"
-        self.diff_from_time = DateTime.create(2024, 6, 15, 18, 35, 20, tz=UTC)
-        self.diff_to_time = DateTime.create(2024, 6, 15, 18, 49, 40, tz=UTC)
-
-    def _build_nodes(self, num_nodes: int, num_sub_fields: int) -> set[EnrichedDiffNode]:
-        nodes = {self.build_diff_node(num_sub_fields=num_sub_fields) for _ in range(num_nodes)}
-
-        # need to associate the generated child nodes with the DiffRoot directly
-        # b/c that is how the data will actually be shaped
-        nodes_to_check = list(nodes)
-        all_nodes = set()
-        while len(nodes_to_check) > 0:
-            this_node = nodes_to_check.pop(0)
-            all_nodes.add(this_node)
-            for rel in this_node.relationships:
-                for child_node in rel.nodes:
-                    nodes_to_check.append(child_node)
-        return all_nodes
-
-    async def _save_single_diff(
-        self, diff_repository: DiffRepository, enriched_diff: EnrichedDiffRoot
-    ) -> EnrichedDiffs:
-        base_diff = EnrichedRootFactory.build(
-            base_branch_name=enriched_diff.base_branch_name,
-            diff_branch_name=enriched_diff.base_branch_name,
-            from_time=enriched_diff.from_time,
-            to_time=enriched_diff.to_time,
-            nodes=self._build_nodes(num_nodes=2, num_sub_fields=1),
-            tracking_id=enriched_diff.tracking_id,
-            partner_uuid=enriched_diff.uuid,
-        )
-        enriched_diff.partner_uuid = base_diff.uuid
-        enriched_diffs = EnrichedDiffs(
-            base_branch_name=self.base_branch_name,
-            diff_branch_name=self.diff_branch_name,
-            diff_branch_diff=enriched_diff,
-            base_branch_diff=base_diff,
-        )
-        await diff_repository.save(enriched_diffs=enriched_diffs)
-        return enriched_diffs
 
     async def test_get_non_existent_diff(self, diff_repository: DiffRepository, reset_database):
         right_now = Timestamp()
